+++ tests/test-files/enum/ok.ts
@@ -1,0 +1,5 @@
+enum MyEnum {
+    MY_ENUM_KEY = 5,
+}
+
+void MyEnum.MY_ENUM_KEY;