+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to my ESLint config will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### Fixed
+- No-shadow rule errors when using enums. 
+
 ## [2.0.0] - 2020-09-07
 ### BREAKING CHANGES
 - Changed 'optionalDependencies' to 'peerDependencies.